refactor: MakeDiary
